# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ZGRZYT API",
        "description": "Dokumentacja API dla systemu Zgłoszeń i Rejestru Zdarzeń Technicznych (ZGRZYT). Jest to webowa aplikacja do zgłaszania, obsługi i monitorowania błędów w systemach informatycznych.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/logout": {
            "post": {
                "tags": [
                    "Autoryzacja"
                ],
                "summary": "Wylogowanie użytkownika (unieważnienie tokena API)",
                "description": "Unieważnia token dostępowy API (Bearer token), z którym zostało wykonane żądanie. Jest to metoda wylogowania dla klientów API (np. aplikacji frontendowej Vue).",
                "operationId": "logoutUser",
                "responses": {
                    "200": {
                        "description": "Token pomyślnie unieważniony.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Błąd autoryzacji (brak lub nieprawidłowy token)."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/tickets/{ticket}/messages": {
            "post": {
                "tags": [
                    "Wiadomości"
                ],
                "summary": "Dodaje nową wiadomość do zgłoszenia",
                "description": "Umożliwia zalogowanemu użytkownikowi dodanie nowej wiadomości (komentarza) do istniejącego zgłoszenia. Użytkownik musi mieć uprawnienia do interakcji ze zgłoszeniem.",
                "operationId": "createMessage",
                "parameters": [
                    {
                        "name": "ticket",
                        "in": "path",
                        "description": "ID zgłoszenia",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Treść nowej wiadomości.",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "body"
                                ],
                                "properties": {
                                    "body": {
                                        "type": "string",
                                        "example": "Próbowałem zrestartować komputer, ale to nie pomogło."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Wiadomość pomyślnie dodana.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MessageFull"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Błąd autoryzacji."
                    },
                    "403": {
                        "description": "Brak uprawnień do dodawania wiadomości w tym zgłoszeniu."
                    },
                    "404": {
                        "description": "Nie znaleziono zgłoszenia."
                    },
                    "422": {
                        "description": "Błąd walidacji (np. pusta treść)."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Sesja"
                ],
                "summary": "Wylogowanie użytkownika (sesja webowa)",
                "description": "Kończy sesję użytkownika w aplikacji webowej (unieważnia sesję).",
                "operationId": "logoutSession",
                "responses": {
                    "200": {
                        "description": "Pomyślnie wylogowano"
                    },
                    "401": {
                        "description": "Błąd autoryzacji (użytkownik niezalogowany)"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "User": {
                "title": "User",
                "description": "Model użytkownika",
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true,
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Jan Kowalski"
                    },
                    "login": {
                        "type": "string",
                        "example": "jkowalski"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "j.kowalski@example.com"
                    },
                    "role": {
                        "type": "string",
                        "example": "user",
                        "enum": [
                            "user",
                            "it",
                            "admin"
                        ]
                    },
                    "active": {
                        "description": "Status aktywności konta.",
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    }
                },
                "type": "object"
            },
...
            "Message": {
                "title": "Message",
                "description": "Model wiadomości w zgłoszeniu",
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true,
                        "example": 1
                    },
                    "body": {
                        "type": "string",
                        "example": "Proszę o więcej szczegółów."
                    },
                    "ticket_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "sender_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    }
                },
                "type": "object"
            },
...
            "MessageFull": {
                "title": "MessageFull",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/Message"
                    },
                    {
                        "properties": {
                            "sender": {
                                "$ref": "#/components/schemas/User"
                            }
                        },
                        "type": "object"
                    }
                ]
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}